PersistentStateを別ファイルに分離します
