+++ src/useState.ts
@@ -1,5 +1,4 @@
 import { map, action } from "nanostores";
-import { persistentAtom } from "@nanostores/persistent";
 
 interface GameStore {
   scenarioIndex: number;
@@ -8,35 +7,6 @@ interface GameStore {
 export const store = map<GameStore>({
   scenarioIndex: -1,
 });
-
-export interface PersistentGameStore {
-  part: string;
-  chapter: number;
-  monologue1AlreadyRead: boolean;
-  monologue2AlreadyRead: boolean;
-}
-
-export const persistentStore = persistentAtom<PersistentGameStore>(
-  "game",
-  {
-    part: "monologue1",
-    chapter: 0,
-    monologue1AlreadyRead: false,
-    monologue2AlreadyRead: false,
-  },
-  {
-    encode: JSON.stringify,
-    decode: JSON.parse,
-  }
-);
-
-export const update = (payload: Partial<PersistentGameStore>): void => {
-  const t = persistentStore.get();
-  persistentStore.set({
-    ...t,
-    ...payload,
-  });
-};
 
 export const resetCounter = action(store, "resetCounter", (store) => {
   store.set({
